+++ CHANGES.md
@@ -32,7 +32,7 @@ We adhere to the [keepachangelog](https://keepachangelog.com/en/1.0.0/) format (
 ## [4.0.0] - 2025-09-24
 ### Changes
 * **BREAKING** Bump the required Java to `17`. ([#2375](https://github.com/diffplug/spotless/issues/2375), [#2540](https://github.com/diffplug/spotless/pull/2540))
-* **BREAKING** Renamed `RemoveWildcardImportsStep` to `ForbidWildcardImportsStep`. ([#2633](https://github.com/diffplug/spotless/pull/2633))
+* **BREAKING** Renamed `RemoveWildcardImportsStep` to `ExpandWildcardImportsStep`. ([#2633](https://github.com/diffplug/spotless/pull/2633))
 * Bump JGit from `6.10.1` to `7.3.0` ([#2257](https://github.com/diffplug/spotless/pull/2257))
   * Adds support for worktrees (fixes [#1765](https://github.com/diffplug/spotless/issues/1765))
 * Bump default `google-java-format` version to latest `1.24.0` -> `1.28.0`. ([#2345](https://github.com/diffplug/spotless/pull/2345))